refactor: 월을 계산하는 로직 분리
